organize task add patches
